backend/requirements.txt
Django==5.0.4
asgiref==3.8.1

django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1

pillow==10.4.0
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
whitenoise==6.6.0
gunicorn==23.0.0

dj-database-url==2.1.0
psycopg2-binary==2.9.11
requests>=2.31.0,<3.0
